requirements.txt
numpy
pycairo
moviepy >= 1.0
# matplotlib
# jupyter
# PyQt5